requirements.txt
albumentations==0.5.2
h5py==2.10.0
matplotlib==3.1.2
numpy==1.17.4
pandas==0.25.3
Pillow==7.0.0
prettytable==2.1.0
scikit-image==0.16.2
scikit-learn==0.24.1
seaborn==0.11.1
tensorboardX==2.1
torch==1.8.0+cu111
torchaudio==0.8.0
torchvision==0.9.0+cu111
tqdm==4.59.0
python==3.8.5
